Move drawing utils to utils folder
Move drawing utils to utils folder
